requirements.txt
fastapi
uvicorn[standard]
pyproj
rasterio
pandas
geopandas
pillow
requests
openeo
openmeteo_requests
requests_cache
retry_requests
h5py
scikit-learn
tables
pyarrow
netCDF4
h5netcdf
ipython
